Move errors into single file
