+++ src/lib/seam/connect/index.ts
@@ -1,6 +1,5 @@
-export * from './api-error.js'
-export * from './invalid-input-error.js'
 export * from './options.js'
 export * from './routes/index.js'
 export * from './seam-http.js'
+export * from './seam-http-error.js'
 export * from 'lib/params-serializer.js'
